- Add VCS_ENABLE_LINT option for running LINT in QA simulation builds.
- Driver appears to work now on real hardware after many changes.
  Rewrote the scoreboard in SystemVerilog.  Lots more cleanup and performance
  work still pending.

+++ modules/bluespec/common/fpgaenv/physical-platform/physical-devices/qa/polled-driver/AFU.h
@@ -44,7 +44,7 @@ public:
     return *(volatile uint32_t *)(dsm_buffer->virtual_address + offset);
   }
 
-  inline volatile uint64_t read_dsm_64(uint64_t offset) {
+  inline volatile uint64_t read_dsm_64(uint32_t offset) {
     return *(volatile uint64_t *)(dsm_buffer->virtual_address + offset);
   }
